# crates/wormhole-cli/tests/fixtures/local-api.openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Wormhole Local API",
    "description": "Control the per-user Wormhole daemon, manage exposures, inspect requests, and diagnose local providers. Operational routes require the bearer token stored in the private runtime directory.",
    "license": {
      "name": "MIT",
      "identifier": "MIT"
    },
    "version": "0.0.0"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:52731",
      "description": "Local Wormhole daemon"
    }
  ],
  "paths": {
...
    "/v1/endpoints": {
      "get": {
        "tags": [
          "Endpoints"
        ],
        "summary": "List active endpoints",
        "operationId": "endpoints",
        "parameters": [
          {
            "name": "service",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActiveEndpoint"
                  }
                }
              }
            }
          }
        }
      }
    },
...
    "/v1/requests": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "List captured requests",
        "operationId": "requests",
        "parameters": [
          {
            "name": "endpoint",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int32",
              "minimum": 0
            }
          },
          {
            "name": "since",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/CapturedRequest"
                  }
                }
              }
            }
          }
        }
      },
      "delete": {
        "tags": [
          "Requests"
        ],
        "summary": "Clear captured requests",
        "operationId": "clear_requests",
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ClosedResponse"
                }
              }
            }
          }
        }
      }
    },
    "/v1/requests/{id}": {
      "get": {
        "tags": [
          "Requests"
        ],
        "summary": "Get a captured request",
        "operationId": "request",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CapturedRequest"
                }
              }
            }
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/v1/requests/{id}/replay": {
      "post": {
        "tags": [
          "Requests"
        ],
        "summary": "Replay a captured request",
        "operationId": "replay",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ReplayResponse"
                }
              }
            }
          },
          "400": {
            "description": ""
          },
          "404": {
            "description": ""
          }
        }
      }
    },
    "/v1/services": {
      "get": {
        "tags": [
          "Services"
        ],
        "summary": "List services",
        "operationId": "services",
        "parameters": [
          {
            "name": "watch",
            "in": "query",
            "description": "Wait for the next status change",
            "required": false,
            "schema": {
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ServiceView"
                  }
                }
              }
            }
          },
          "401": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Services"
        ],
        "summary": "Create a service",
        "operationId": "create_service",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateServiceRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActiveEndpoint"
                  }
                }
              }
            }
          },
          "207": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/ActiveEndpoint"
                  }
                }
              }
            }
          },
          "409": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          },
          "502": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiErrorBody"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ActiveEndpoint": {
        "type": "object",
        "description": "A live or recently stopped exposure.",
        "required": [
          "id",
          "service",
          "driver",
          "urls",
          "status",
          "since"
        ],
        "properties": {
          "buffered_delivered": {
            "type": "integer",
            "format": "int64",
            "description": "Number of buffered webhooks delivered in this daemon lifetime.",
            "minimum": 0
          },
          "buffered_failed": {
            "type": "integer",
            "format": "int32",
            "description": "Number quarantined at the relay when last reported.",
            "minimum": 0
          },
          "buffered_pending": {
            "type": "integer",
            "format": "int32",
            "description": "Number waiting at the relay when last reported.",
            "minimum": 0
          },
          "driver": {
            "type": "string",
            "description": "Driver registry name."
          },
          "id": {
            "type": "string",
            "format": "uuid",
            "description": "Manager-local endpoint identifier."
          },
          "service": {
            "type": "string",
            "description": "Owning service name."
          },
          "since": {
            "type": "string",
            "format": "date-time",
            "description": "Time this status record was created."
          },
          "status": {
            "$ref": "#/components/schemas/EndpointStatus",
            "description": "Current lifecycle status."
          },
          "urls": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "description": "Public URLs currently assigned."
          }
        }
      },
      "ApiErrorBody": {
        "type": "object",
        "required": [
          "error"
        ],
        "properties": {
          "error": {
            "$ref": "#/components/schemas/ApiErrorDetail"
          }
        }
      },
      "ApiErrorDetail": {
        "type": "object",
        "required": [
          "code",
          "message"
        ],
        "properties": {
          "code": {
            "type": "string"
          },
          "message": {
            "type": "string"
          }
        }
      },
      "BufferPolicy": {
        "type": "object",
        "description": "Limits applied to offline webhook buffering.",
        "required": [
          "max_requests",
          "max_body_bytes",
          "ttl_secs"
        ],
        "properties": {
          "max_body_bytes": {
            "type": "integer",
            "format": "int64",
            "description": "Maximum body bytes retained per request.",
            "minimum": 0
          },
          "max_requests": {
            "type": "integer",
            "format": "int32",
            "description": "Maximum buffered request count.",
            "minimum": 0
          },
          "ttl_secs": {
            "type": "integer",
            "format": "int64",
            "description": "Maximum retention time in seconds.",
            "minimum": 0
          }
        }
      },
      "CapturedHeader": {
        "type": "object",
        "description": "One redacted captured HTTP header.",
        "required": [
          "name",
          "value_b64"
        ],
        "properties": {
          "name": {
            "type": "string"
          },
          "value_b64": {
            "type": "string"
          }
        }
      },
      "CapturedRequest": {
        "type": "object",
        "description": "Memory-only captured HTTP exchange.",
        "required": [
          "id",
          "bind_id",
          "method",
          "uri",
          "headers",
          "body",
          "body_truncated",
          "response_headers",
          "response_body_prefix",
          "response_body_truncated",
          "duration_ms",
          "delivery",
          "captured_at"
        ],
        "properties": {
          "bind_id": {
            "type": "string",
            "format": "uuid"
          },
          "body": {
            "type": "string"
          },
          "body_truncated": {
            "type": "boolean"
          },
          "captured_at": {
            "type": "string",
            "format": "date-time"
          },
          "delivery": {
            "type": "string"
          },
          "duration_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "endpoint_id": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid"
          },
          "headers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CapturedHeader"
            }
          },
          "id": {
            "type": "string",
            "format": "uuid"
          },
          "method": {
            "type": "string"
          },
          "response_body_prefix": {
            "type": "string"
          },
          "response_body_truncated": {
            "type": "boolean"
          },
          "response_headers": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/CapturedHeader"
            }
          },
          "response_status": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "minimum": 0
          },
          "uri": {
            "type": "string"
          }
        }
      },
      "ClosedResponse": {
        "type": "object",
        "required": [
          "closed"
        ],
        "properties": {
          "closed": {
            "type": "boolean"
          }
        }
      },
      "CreateServiceRequest": {
        "type": "object",
        "required": [
          "service",
          "endpoints"
        ],
        "properties": {
          "default_remote": {
            "type": [
              "string",
              "null"
            ]
          },
          "endpoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/EndpointSpec"
            }
          },
          "project_id": {
            "type": [
              "string",
              "null"
            ]
          },
          "remotes": {
            "type": [
              "object",
              "null"
            ],
            "additionalProperties": {
              "$ref": "#/components/schemas/Remote"
            },
            "propertyNames": {
              "type": "string"
            }
          },
          "service": {
            "$ref": "#/components/schemas/Service"
          }
        }
      },
...
      "EdgeAuth": {
        "type": "object",
        "description": "Access-control material enforced at a public HTTP edge.",
        "properties": {
          "basic": {
            "type": [
              "string",
              "null"
            ],
            "description": "Basic-auth credential in `user:password` form."
          },
          "bearer": {
            "type": [
              "string",
              "null"
            ],
            "description": "Bearer token expected in the Authorization header."
          },
          "link_key": {
            "type": [
              "string",
              "null"
            ],
            "description": "Base64-encoded HMAC key for expiring share links."
          }
        }
      },
      "EndpointSpec": {
        "type": "object",
        "description": "One desired public exposure through one driver instance.",
        "required": [
          "proto",
          "driver",
          "persist",
          "inspect"
        ],
        "properties": {
          "auth": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/EdgeAuth",
                "description": "Relay-edge authentication policy."
              }
            ]
          },
          "auto_host": {
            "type": "boolean",
            "description": "Permit the relay to suffix an automatically generated host on conflict."
          },
          "buffer": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/BufferPolicy",
                "description": "Offline HTTP buffering policy."
              }
            ]
          },
          "capture_body_max": {
            "type": "integer",
            "format": "int64",
            "description": "Maximum complete request body retained for inspection.",
            "minimum": 0
          },
          "domain": {
            "type": [
              "string",
              "null"
            ],
            "description": "Requested offered server domain."
          },
          "driver": {
            "type": "string",
            "description": "Registered driver name."
          },
          "host": {
            "type": [
              "string",
              "null"
            ],
            "description": "Requested server-owned subdomain label."
          },
          "inspect": {
            "type": "boolean",
            "description": "Capture requests for inspection."
          },
          "inspect_assets": {
            "type": "boolean",
            "description": "Include static asset requests in inspection capture."
          },
          "persist": {
            "$ref": "#/components/schemas/Persistence",
            "description": "Endpoint lifetime."
          },
          "proto": {
            "$ref": "#/components/schemas/ServiceProto",
            "description": "Public protocol, overwritten from the service by the manager."
          },
          "public_port": {
            "type": [
              "integer",
              "null"
            ],
            "format": "int32",
            "description": "Provider-side public port.",
            "minimum": 0
          },
          "qualifier": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional provider mode."
          },
          "remote": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional named Wormhole remote."
          },
          "reservation": {
            "type": [
              "string",
              "null"
            ],
            "format": "uuid",
            "description": "Reservation used to reclaim a persistent Wormhole bind."
          },
          "retry": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/RetryPolicy",
                "description": "Local delivery retries."
              }
            ]
          }
        }
      },
      "EndpointStatus": {
        "oneOf": [
          {
            "type": "string",
            "description": "Public routing is ready.",
            "enum": [
              "online"
            ]
          },
          {
            "type": "string",
            "description": "Driver is reconnecting.",
            "enum": [
              "reconnecting"
            ]
          },
          {
            "type": "string",
            "description": "Endpoint is intentionally stopped.",
            "enum": [
              "offline"
            ]
          },
          {
            "type": "object",
            "description": "Driver terminated with an error.",
            "required": [
              "error"
            ],
            "properties": {
              "error": {
                "type": "string",
                "description": "Driver terminated with an error."
              }
            }
          }
        ],
        "description": "Current endpoint lifecycle status."
      },
...
      "Persistence": {
        "type": "string",
        "description": "Lifetime of a bind reservation.",
        "enum": [
          "temporary",
          "persistent"
        ]
      },
...
      "Remote": {
        "type": "object",
        "description": "One named Wormhole relay.",
        "required": [
          "addr",
          "server_name"
        ],
        "properties": {
          "addr": {
            "type": "string",
            "description": "UDP authority of the relay."
          },
          "https_addr": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional HTTPS authority for WebSocket fallback; defaults to `server_name:443`."
          },
          "identity": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional per-remote identity override."
          },
          "server_name": {
            "type": "string",
            "description": "TLS and handshake-bound server name."
          },
          "transport": {
            "$ref": "#/components/schemas/Transport",
            "description": "Preferred control transport."
          },
          "trusted_ca": {
            "type": [
              "string",
              "null"
            ],
            "description": "Optional exclusive development CA root."
          }
        }
      },
...
      "ReplayResponse": {
        "type": "object",
        "required": [
          "status",
          "duration_ms"
        ],
        "properties": {
          "duration_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          }
        }
      },
      "RetryPolicy": {
        "type": "object",
        "description": "Local HTTP delivery retry policy reserved for Stage 07 behavior.",
        "required": [
          "max_attempts",
          "initial_delay_ms"
        ],
        "properties": {
          "initial_delay_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "max_attempts": {
            "type": "integer",
            "format": "int32",
            "minimum": 0
          },
          "max_body_bytes": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "max_delay_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          },
          "retry_5xx": {
            "type": "boolean"
          },
          "retry_connect": {
            "type": "boolean"
          },
          "total_deadline_ms": {
            "type": "integer",
            "format": "int64",
            "minimum": 0
          }
        }
      },
      "Service": {
        "type": "object",
        "description": "A local thing to expose.",
        "required": [
          "name",
          "target",
          "proto"
        ],
        "properties": {
          "name": {
            "type": "string",
            "description": "Stable local service name."
          },
          "proto": {
            "$ref": "#/components/schemas/ServiceProto",
            "description": "Local delivery protocol."
          },
          "target": {
            "$ref": "#/components/schemas/TargetSchema",
            "description": "Unresolved local destination."
          }
        }
      },
      "ServiceProto": {
        "type": "string",
        "description": "Protocol spoken by a local service.",
        "enum": [
          "http",
          "tcp"
        ]
      },
      "ServiceView": {
        "type": "object",
        "required": [
          "project_id",
          "service",
          "endpoints"
        ],
        "properties": {
          "endpoints": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ActiveEndpoint"
            }
          },
          "project_id": {
            "type": "string"
          },
          "service": {
            "$ref": "#/components/schemas/Service"
          }
        }
      },
...
      "TargetSchema": {
        "oneOf": [
          {
            "$ref": "#/components/schemas/PortTargetSchema"
          },
          {
            "$ref": "#/components/schemas/HostPortTargetSchema"
          },
          {
            "$ref": "#/components/schemas/IfaceTargetSchema"
          }
        ],
        "description": "Schema-only representation of the tagged target wire contract."
      },
      "Transport": {
        "type": "string",
        "enum": [
          "auto",
          "quic",
          "ws"
        ]
      }
    },
    "securitySchemes": {
      "bearer_auth": {
        "type": "http",
        "scheme": "bearer"
      }
    }
  },
  "security": [
    {
      "bearer_auth": []
    }
  ],
...
}